continue logging to disk if interrupted (wake-lock)
When the app goes out of foreground while logging to disk, e.g.,
because the screen turns off or another app is started, acquire partial
wake-lock and continue logging.

If not logging to disk, behavior is unchanged, i.e., updates are
suspended.

This feature is (particularly) experimental and definitely needs
testing and/or a review.

See #1